# docs/openapi.yaml
openapi: 3.0.0
info:
  title: SiTukang API
  description: |
    Complete API Documentation for SiTukang - A Single Backend for Customers and Workers.
    Authentication is handled via RBAC with JWT access tokens.
  version: 1.0.0
servers:
  - url: https://xryz-gcw-situkang.hf.space/v1
    description: Production Server (Hugging Face)
  - url: http://localhost:8080/v1
    description: Local Development Server
security:
  - BearerAuth: []
paths:
...
  /payments/sandbox-callback:
    post:
      tags: [Sandbox Payment Gateway]
      summary: Sandbox Payment Gateway Callback
      description: Webhook simulation that instantly processes checkout status transitions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_id, status]
              properties:
                payment_id: { type: string, example: "9f9a4a7b-3b3d-4c3e-8c3d-3b3d4c3e8c3d" }
                status: { type: string, enum: [success, failed], example: success }
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: success }
                  message: { type: string, example: "Pembayaran sandbox sukses diproses" }
                  data:
                    type: object
                    properties:
                      payment_id: { type: string, example: "9f9a4a7b-3b3d-4c3e-8c3d-3b3d4c3e8c3d" }
                      order_id: { type: string, example: "8e8a4a7b-2b2d-3c3e-7c3d-2b2d3c3e7c3d" }
                      payment_status: { type: string, example: "paid" }
                      paid_at: { type: string, format: date-time, example: "2026-05-30T10:15:45Z" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token containing user/worker claims.
...